chore(test-requirements.txt): add nose-ignore-docstring
Our pep257 checker requires every public method to have a docstring, but
the problem is, nosetests will use the first line of the docstring in
the test name output. This is kind of a drag, because in verbose test
output, you want to see the test module path and name, not the docstring
header.

Adding this plugin for nose fixes that.