test(test_schema): add tests for `schema` module
Add tests for schema definitions and associated utility functions.

More tests need to be added to check additional realistic
Checkmatefiles, but this is a good start. This set of tests exercises
tests validation of not only a valid Checkmatefile, but also tests
validation and the error checking/reporting for a Checkmatefile with
some errors in it.