fix(schema): fix coercion errors
Voluptuous expects all schema callables to act
like a type() call; i.e. it should return the
requested type, but not modify the supplied
arguments. Voluptuous then determines if it should
coerce or not (unless Coerce is used explicitly).

Many of our schema callables were not handling
this correctly and resulted in recursion and other
errors. These changes address that.