ref(topology): behave closer to a dict
Topology() should behave like a dict as much as
possible. Properties and methods are there for
convenience and readability. This lets the client
of this module build a topology incrementally
starting with a blank dict and decide if they want
to use .services, item getter, .setdefault, etc…

No assumptions are made up front and no
constructor override is needed.